Hold text in UTF-8 internally, and convert on the way to the wire
GoldED+ can now keep message text, headers and the screen in UTF-8
rather than in a single-byte charset. Set XLATLOCALSET to UTF-8 to use
it. This is about the operating system - console, filenames, editing -
and not about what goes into the message base: the transport charset
stays whatever XLATEXPORT says, so messages are still written as CP866
unless told otherwise, and a UTF-8 session announces them as CHRS
level 4 per FSC-0054.

Charset conversion is rewritten and lives in one place,
goldlib/gall/grecode.cpp. It uses, in this order, whichever is
available: iconv; the operating system's own converter - the Win32
codepage API on Windows, ULS on OS/2; and only where neither exists the
compiled-in tables. goldlib/gall/gutf8.cpp holds the primitives
underneath - decoding, encoding, stepping a character at a time, and
measuring how many screen columns a string occupies.

Text is measured in columns rather than bytes wherever the two differ,
and a field that has to be cut is cut between characters, whether it is
being fitted to the screen or into the fixed-width header fields of a
message base.

New configuration keywords:

  XLATCONFIGSET  the charset GoldED+'s own files are in - golded.cfg
                 and the language file with it, so a CP866 goldlang
                 can be read in a UTF-8 session.
  XLATAREASET    the charset area descriptions are in; they usually
                 come from an area file the tosser writes and need not
                 match. Falls back to XLATCONFIGSET.

golded.cfg is now looked for in a defined order - the binary's own
directory, the current directory, the user's configuration directories,
then the system ones - rather than in a single place fixed at build
time.

Fixes that stand on their own:

 - a message list showed the subject in the wrong charset whenever the
   message did not use the one the area imports. A CHRS kludge belongs
   to the message text and the list does not read the text; but JAM,
   Squish, SMB and *.MSG can all produce it from the header for very
   little, and now do. Hudson and the rest keep the old behaviour,
   deliberately: there it would cost a seek and a read per line.
 - a search for two words separated by a space found nothing.
 - Backspace in an input field deleted a whole word rather than a
   character, and Ctrl-Backspace did the opposite.
 - MIME encoded-words in a header were decoded twice when the line
   parser had already converted the text.
 - quote colouring looked random in some messages: the UUE detector
   took any pure-ASCII quote of 42 to 62 bytes for encoded data.
 - the Windows clipboard went through the OEM codepage in both
   directions.
 - CWSDPMI is packed with the DOS distribution again; without it the
   archive held binaries that will not start on a bare DOS.
 - `make strip' ran the host's strip, which cannot read a cross-built
   binary.
 - dist-sx86.sh lost the date out of the package name on Solaris.
 - the 32-bit DOS builds handed int386x() and int86x() uninitialised
   segment registers.
 - two pointer differences were taken through a long, which is narrower
   than a pointer on 64-bit Windows.

CPU identification is rewritten over <cpuid.h>/__cpuidex() instead of
32-bit inline assembler, and decodes the extended family and model, so
modern processors are no longer reported as an Athlon 64.

Build:

 - CMake finds curses and iconv by probing rather than by assuming, and
   prefers a package manager's copy with the system's own behind it.
 - the GNU makefiles get the same treatment, and the historical targets
   keep working: DOS/DJGPP, Cygwin, OS/2, BeOS, Solaris, the BSDs.
 - new gedbcc.mak (Borland C++ 5.02/5.5.1) and gedwcc.mak (Open Watcom,
   four targets).
 - OS/2 builds with either toolchain and can be cross-built off the
   machine; charset conversion there goes through ULS.
 - the DJGPP build can recode through iconv, which leaves nothing on
   the compiled-in tables.
 - new dist-hpkg.sh and docs/haiku.txt for a Haiku package.

Every target above was built and started, and the ones with a message
base driven through real areas - the platforms and toolchains are
listed in docs/notework.txt, along with the one that builds and starts
but does not run: Borland's 32-bit DOS target, which faults inside the
extender's own runtime.

+++ golded3/getpls.cpp
@@ -816,7 +816,7 @@ int TemplateToText(int mode, GMsg* msg, GMsg* oldmsg, const char* tpl, int origa
                             }
                         }
                         ptr = strskip_wht(quotestr);
-                        y = (int)((long)ptr-(long)quotestr);
+                        y = (int)(ptr-quotestr);
                         n = 0;
                         *buf = NUL;
                         while(oldmsg->line[n])
@@ -864,7 +864,7 @@ int TemplateToText(int mode, GMsg* msg, GMsg* oldmsg, const char* tpl, int origa
                                         ptr--;
                                     }
                                 }
-                                x = (int)((long)ptr-(long)qbuf);
+                                x = (int)(ptr-qbuf);
                                 sprintf(buf, "%*.*s%*.*s>%s %s",
                                         y, y, quotestr, x, x, qbuf, qbuf+x, quote
                                        );
